Renames based on defacto standards
